refactor deposit form
